new try

+++ sudoku.c
@@ -51,7 +51,7 @@ int is_valid(Node* n){
 
 List* get_adj_nodes(Node* n){
   List* list=createList();
-  unsigned long i,j;
+  int i,j;
   return list;
 }
